Major revision
updated several files
