+++ Light.h
@@ -1,22 +1,28 @@
-/* high-level lights API
-* --
-* Contains "lights" definition and functions. Lights
-* are an abstraction of pinball real hardware "lamps".
+/* *** Gottileb System 80/B SIM PRB ***
+* (SIMulation Pinball Replacement control Board)
+* software for Teensy 3.x board developed with Arduino IDE
+* under GNU General Public Livence v3
+* ---
+* LIGHT OBJECT DEFINITION
+* ---
+* "Lights" are an abstraction of pinball real hardware "lamps".
 * In order to make a light blink of pulse you have to
-* call steadily the function update() with a convenient
-* period.
+* call steadily update() function with a convenient time interval.
+* Contains C++ code.
 */
+
+#ifndef Light_h
+#define Light_h
 
 #define DEF_BLNK_PERIOD 200 // default light blink rate
 #define DEF_PULSE_PERIOD 200 // default light pulse rate
 #define LSTAGE_SIZE 48 // stage total lights count
 
 // defines different possible light states
-enum lightState { 
+enum lightState {
   OFF_L,    // light off
   ON_L,     // light on
   BLINK_L,  // alternate on/off ligth state (with duty cycle)
-  PULSE_L   // temporary light state inversion -> (then) ON or OFF
   };
 
 // software mid-level light object, with its own state
@@ -31,27 +37,31 @@ class Light {
     uint32_t activationTime; // time when lamp has activated (from OFF to another state)
     uint16_t activePeriod; // time limit for any state different than OFF_L (ms)
         // (0 = unlimited time)
-    
+    bool impulse; // temporary inverted on/off state
+    uint32_t pulseTime; // time when lamp pulse has activated
+    uint16_t pulsePeriod; // pulse state duration
+
     Light();
     void copy(Light &from);
     void reset();
     inline void set(lightState st);
-    inline void blink(lightState st, uint16_t blinkP);
-    void blink(lightState st, uint16_t blinkP, byte blinkDutyC);
+    inline void set(lightState st, uint16_t blinkP);
+    inline void blink(uint16_t blinkP, byte ticks);
+    void set(lightState st, uint16_t blinkP, byte blinkDutyC);
     void invert();
     void pulse(uint16_t pulseP);
-    void blink(uint16_t blinkP, byte ticks);
     void update(uint32_t ms);
     inline bool isActive() { return (state != OFF_L); }
 };
 
 // constructor with default property values
-Light::Light() { 
+Light::Light() {
   on = false;
+  impulse = false;
   state = OFF_L;
   blinkPeriod = 500;
-  blinkDutyCycle = 50; 
-  activePeriod = 0; 
+  blinkDutyCycle = 50;
+  activePeriod = 0;
 }
 
 // convenient funcion for copying a Light state into another
@@ -82,11 +92,11 @@ inline void Light::set(lightState st) {
 // define the state of the light
 // st: light state
 // blinkP: blink period (on + off) [milliseconds]
-inline void Light::blink(lightState st, uint16_t blinkP) {
+inline void Light::set(lightState st, uint16_t blinkP) {
   set(st, blinkP, 50);
 }
 
-void Light::blink(lightState st, uint16_t blinkP, byte blinkDutyC) {
+void Light::set(lightState st, uint16_t blinkP, byte blinkDutyC) {
   switch (st) {
     case OFF_L:
       on = false;
@@ -98,57 +108,61 @@ void Light::blink(lightState st, uint16_t blinkP, byte blinkDutyC) {
       if (blinkP == 0) state = ON_L;
       on = true;
       break;
-    case PULSE_L:
-      on = !on;
-      break;
     default:
       return;
   }
   state = st;
-  if (st != OFF_L) activationTime = millis();
+  if (on) activationTime = millis();
   blinkPeriod = blinkP;
-  if (blinkDutyC == 0) blinkDutyCycle = 50; 
+  if (blinkDutyC == 0) blinkDutyCycle = 50;
   else blinkDutyCycle = blinkDutyC;
   activePeriod = 0; // unlimited state time (default)
 }
 
-// inverts light ON/OFF status
+inline void Light::blink(uint16_t blinkP, byte ticks) {
+  set(BLINK_L, blinkP, 50);
+  activePeriod = blinkP * ticks;
+}
+
+// inverts light ON/OFF state
 void Light::invert() {
   if (state == ON_L) state = OFF_L;
   else if (state == OFF_L) state = ON_L;
 }
 
 void Light::pulse(uint16_t pulseP) {
-  set(PULSE_L, (pulseP == 0) ? DEF_PULSE_PERIOD : pulseP);
-}
-
-void Light::blink(uint16_t blinkP, byte ticks) {
-  set(BLINK_L, blinkP, 50);
-  activePeriod = blinkP * ticks;
+  pulsePeriod = pulseP;
+  impulse = true;
+  pulseTime = millis();
+  on = !on;
 }
 
 void Light::update(uint32_t ms) {
-  uint32_t bt; // blink ON time
-    
+  uint32_t t; // blink ON time
+
+  // blink mode
+  if (state == BLINK_L) {
+    t = (ms - activationTime) % (uint32_t)blinkPeriod;
+    on = ((t*100 / blinkPeriod) < blinkDutyCycle);
+  }
+
+  // pulse state
+  if (impulse) {
+    if ((ms - pulseTime) > pulsePeriod) {
+      impulse = false;
+      on = !on;
+    }
+  }
+
   // checks light active time
-  if (state != OFF_L) {
-    if (activePeriod > 0 && 
-      (ms - activationTime) >= (uint32_t)activePeriod) {
+  if (on) {
+    if (activePeriod > 0 && (ms - activationTime) >= (uint32_t)activePeriod) {
       activePeriod = 0; // resets previous active-time setting
       state = OFF_L;
     }
   }
   
-  // implements special light modes
-  if (state == BLINK_L) {
-    bt = (ms - activationTime) % (uint32_t)blinkPeriod;
-    on = ((bt*100 / blinkPeriod) < blinkDutyCycle);
-  }
-  else if (state == PULSE_L) {
-    if ((ms - activationTime) >= blinkPeriod) {
-      on = !on;
-      if (state) state = ON_L; else state = OFF_L; // auto state switch
-    }
-  }
 }
+
+#endif
