+++ flashData.h
@@ -1,3 +1,13 @@
+/* *** Gottileb System 80/B SIM PRB ***
+* (SIMulation Pinball Replacement control Board)
+* software for Teensy 3.x board developed with Arduino IDE
+* under GNU General Public Livence v3
+* ---
+* FLASH MEMORY DATA I/O
+* ---
+* Functions for pinball non-volatile data retrieve/store.
+*/
+
 #include <EEPROM.h>
 #include "sdio.h"
 
@@ -14,9 +24,9 @@ PROGMEM const byte factorySwSettings[] = { // Factory "Op-Adj" switch settings v
 
 PROGMEM const byte factoryStatsBytes[] = { // Factory stats values
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0,00, 0x40, 0x4b, 0x4c, 0x00, 0x80, 0x96, 0x98, 0x00, 0x00, 0x2d, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0,00, 0x40, 0x4b, 0x4c, 0x00, 0x80, 0x96, 0x98, 0x00, 0x00, 0x2d,
   0x31, 0x01, 0x00, 0x00, 0x60, 0xe3, 0x16, 0x00, 0x52, 0x41, 0x4d, 0x00, 0x80, 0x4f, 0x12, 0x00,
-  0x52, 0x41, 0x4d, 0x00, 0x40, 0x42, 0x0f, 0x00, 0x52, 0x41, 0x4d, 0x00, 0x00, 0x35, 0x0c, 0x00,  
+  0x52, 0x41, 0x4d, 0x00, 0x40, 0x42, 0x0f, 0x00, 0x52, 0x41, 0x4d, 0x00, 0x00, 0x35, 0x0c, 0x00,
   0x52, 0x41, 0x4d, 0x00, 0x20, 0xa1, 0x07, 0x00, 0x52, 0x41, 0x4d, 0x00
 };
 
@@ -43,7 +53,7 @@ byte getSettingBit(byte bitPos) {
   if (bitPos > 31) return false;
   _byte = bitPos >> 3;
   _bit = bitPos % 8;
-  return (_getFVal(_byte) & (1<<_bit)) ? 1 : 0; 
+  return (_getFVal(_byte) & (1<<_bit)) ? 1 : 0;
 }
 
 bool saveSettingByte(byte addr, byte byteMask, byte value) {
@@ -110,4 +120,3 @@ uint32_t getStat32(byte addr) {
   v |= _getFVal(addr+3);
   return v;
 }
-
