+++ pinGeneric.h
@@ -1,11 +1,21 @@
-#include "tableItems.h"
+/* *** Gottileb System 80/B SIM PRB ***
+* (SIMulation Pinball Replacement control Board)
+* software for Teensy 3.x board developed with Arduino IDE
+* under GNU General Public Livence v3
+* ---
+* PINBALL GENERIC CODE
+* ---
+* Game generic code for basic machine functions,
+* fundamental for all games.
+*/
+
 #include "ledGrid.h"
 #include "baseAPI.h"
 #include "flashData.h"
 #include "displFX.h"
-#include "Light.h"
-#include "LightStage.h"
+#include "LightSet.h"
 #include "ui_menu.h"
+#include "timerTask.h"
 
 #define MAX_PLAYERS 4
 #define TOP_SCORES  5
@@ -23,6 +33,12 @@
 #define setQrelay(s) setLamp(0,s)
 #define setTrelay(s) setLamp(1,s)
 #define setArelay(s) setLamp(14,s)
+
+// standard "lamp" solenoids (controlled as lamps)
+const int RELAY_Q_SOL = 16;
+const int RELAY_T_SOL = 17; 
+const int BRELEASE_SOL = 18;
+const int SND16_SOL = 20;
 
 enum pinball_mode { START_MODE, ATTRACT_MODE, GAME_MODE, TEST_MODE, BOOKKEEP_MODE };
 enum coinChutes { LEFT_CHUTE, CENTER_CHUTE, RIGHT_CHUTE };
@@ -67,8 +83,7 @@ uint16_t switches;
 bool tiltEnabled;
 Light light[LAMPS_COUNT];
 LightStage *lightStage;
-
-extern bool lightsModeEnabled; // lightFX.h
+bool lightStageEnabled;
 
 void initPinball_generic();
 void initGame_generic();
@@ -92,6 +107,7 @@ void resetTilt();
 bool releaseBall();
 
 extern void onEvent(byte sw);
+extern String getGameName();
 
 void initPinball_generic() {
   byte i;
@@ -108,7 +124,7 @@ void initPinball_generic() {
   defaultBallsPerPlay = (getSettingSwitch(25) == 1) ? 3 : 5;
   for (i=0; i<3; i++) coins[i] = 0; // coins reset
   for (i=0; i<MAX_PLAYERS; i++) player[i].score = 0; // scores reset
-  lightsModeEnabled = true;
+  lightStageEnabled = true;
   setPinMode(START_MODE);
   setSolenoid(RELAY_Q_SOL, true); // game over
 }
@@ -121,12 +137,14 @@ void initGame_generic() {
 }
 
 void setPinMode(byte mode) {
+  String s;
+  
   pinballMode = mode;
   modeStep = 0;
   switch (pinballMode) {
     case START_MODE:
       setDisplayText(0, (const char *) F("SYS80/B PRB 1.0 BY FABVOLPI@GMAIL.COM"));
-      String s = new String(F("PRESS START FOR *") + String(GAMENAME) + "*");
+      s = String("PRESS START FOR *" + getGameName() + "*");
       setDisplayText(1, s.c_str());
       break;
     case ATTRACT_MODE:
@@ -150,7 +168,7 @@ void setPinMode(byte mode) {
 void displayScoresAndCredits() {
   String st[4];
   int i, l;
-  
+
   for (i=0; i<4; i++) {
     st[i] = String(zeroStr) + String(player[i].score);
     l = st[i].length();
@@ -163,12 +181,12 @@ void displayScoresAndCredits() {
   writeDisplayText(0, 12, st[1]);
   writeDisplayText(1, st[2]);
   writeDisplayText(1, 12, st[3]);
-  writeDisplayText(1, 9, credits);
+  writeDisplayText(1, 9, String(credits));
 }
 
 void loadHighScores() {
   byte i, j;
-  
+
   for (i=0; i<TOP_SCORES; i++) {
     if (i < 3) {
       j = 36 + i*8;
@@ -192,7 +210,7 @@ void loadAwardScoreLevels() {
 
 void loadCoinsPerCredits() {
   byte i, v;
-  
+
   for (i=0; i<3; i++) {
     v = getSettingByte(i);
     if (v <= 24) {
@@ -219,12 +237,12 @@ void onCoinInserted(byte cch) {
 void incrementPowerOnCounterStat() {
   uint16_t pot;
   byte i;
-  
+
   pot = getStat16(0);
   if (pot == 0) { // on first bootup only
     for (i=0; i<4; i++) saveSettingByte(i, 0xff, factorySwSettings[i]);
     for (i=0; i<76; i++) saveStat8(i, factoryStatsBytes[i]);
-  } 
+  }
   saveStat16(0, ++pot);
 }
 
@@ -262,7 +280,7 @@ void extraBall() {
 
 void addScore(uint32_t sc) {
   byte lev;
-  
+
   player[playerOn].score += sc;
   // check for Score Level award
   lev = player[playerOn].awardLevelReached;
@@ -289,7 +307,7 @@ void addBonus(uint32_t sc) {
 bool switchToNextPlayer() {
   byte prevPlayer;
   bool res;
-  
+
   prevPlayer = playerOn;
   if (++playerOn >= MAX_PLAYERS) playerOn = 0;
   while (player[playerOn].gameOver) {
@@ -322,14 +340,14 @@ void resetTilt() {
 
 bool releaseBall() {
   bool ballReady;
-  
+
   ballReady = (getSwitch(36) > 0);
   if (ballReady) setSolenoid(BRELEASE_SOL, true, 500); // ball release
   return ballReady;
 }
 
 // to call periodically for real lamps update
-void checkLights(uint32_t ms) {
-  lstage->updateLights(ms);
+void updateLamps(uint32_t ms) {
+  if (lightStageEnabled) lightStage->updateLights(ms, true);
   renderLamps();
 }
